feat: Request Sheet View UI
