+++ FITI/FITI/Main/Detail/Request/PickPlace/PickPlaceViewController.swift
@@ -1,5 +1,5 @@
 //
-//  PickDateViewController.swift
+//  PickPlaceViewController.swift
 //  FITI
 //
 //  Created by 홍준혁 on 2023/01/04.
@@ -8,14 +8,16 @@
 import UIKit
 import SnapKit
 import Then
-import FSCalendar
 
-class PickDateViewController: UIViewController {
+class PickPlaceViewController: UIViewController {
+    
+    var isIWillGoSelected : Bool = false
+    var isComeHereSelected : Bool = false
 
     var titleLabel : UILabel = {
         let label = UILabel()
-        label.font = UIFont.systemFont(ofSize: 20.0)
-        label.text = "매칭기간 선택"
+        label.font = UIFont(name: "Avenir-Black", size: 20.0)
+        label.text = "픽업 형태"
         label.textColor = UIColor.customColor(.blue)
         return label
     }()
@@ -37,6 +39,30 @@ class PickDateViewController: UIViewController {
         }
         return view
     }()
+
+    lazy var iWillGoBtn : UIButton = {
+        let btn = UIButton()
+        btn.setImage(UIImage(named: "iWillGoIcon.svg"), for: .normal)
+        btn.addTarget(self, action: #selector(iWillGoEvent), for: .touchUpInside)
+        return btn
+    }()
+    
+    lazy var comeHereBtn : UIButton = {
+        let btn = UIButton()
+        btn.setImage(UIImage(named: "comeHereIcon.svg"), for: .normal)
+        btn.addTarget(self, action: #selector(comeHereEvent), for: .touchUpInside)
+        return btn
+    }()
+    
+    lazy var globalStackView : UIStackView = {
+        let stackView = UIStackView(arrangedSubviews: [iWillGoBtn,comeHereBtn])
+        stackView.axis = .horizontal
+        stackView.spacing = 20
+        stackView.alignment = .bottom
+        stackView.distribution = .fillEqually
+        return stackView
+    }()
+    
     
     private let nextBtn : UIButton = {
         let btn = UIButton()
@@ -51,14 +77,9 @@ class PickDateViewController: UIViewController {
         return btn
     }()
     
-    // calendar
-    let fsCalendar = FSCalendar()
-    
-    
     
     override func viewDidLoad() {
         super.viewDidLoad()
-
         view.backgroundColor = .systemBackground
         
         navigationController?.navigationBar.tintColor = .black
@@ -66,19 +87,16 @@ class PickDateViewController: UIViewController {
         
         self.navigationItem.leftBarButtonItem = UIBarButtonItem(image:UIImage(named: "leftIcon.svg"), style: .plain, target: self, action: #selector(backTapped))
         
-        calendarSetting()
-        
         // Do any additional setup after loading the view.
         setViewHierarchy()
         setConstraints()
-
     }
     
     private func setViewHierarchy() {
         view.addSubview(titleLabel)
         view.addSubview(progressView)
         progressView.addSubview(grayView)
-        view.addSubview(fsCalendar)
+        view.addSubview(globalStackView)
         view.addSubview(nextBtn)
     }
     
@@ -93,14 +111,14 @@ class PickDateViewController: UIViewController {
             make.trailing.equalToSuperview()
         }
         grayView.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
+            make.trailing.equalToSuperview()
             make.width.equalToSuperview().dividedBy(3)
         }
-        fsCalendar.snp.makeConstraints { make in
-            make.top.equalTo(progressView.snp.bottom).offset(10)
-            make.leading.equalToSuperview().offset(30)
-            make.trailing.equalToSuperview().offset(-30)
-            make.bottom.equalToSuperview().offset(-324)
+        globalStackView.snp.makeConstraints { make in
+            make.top.equalTo(progressView.snp.bottom).offset(65)
+            make.leading.equalToSuperview().offset(20)
+            make.trailing.equalToSuperview().offset(-20)
+            make.bottom.equalToSuperview().offset(-430)
         }
         nextBtn.snp.makeConstraints { make in
             make.top.equalToSuperview().offset(770)
@@ -109,28 +127,34 @@ class PickDateViewController: UIViewController {
             make.bottom.equalToSuperview().offset(-40)
         }
     }
-    
-    @objc func nextEvent(){
-        let nextVC = PickPlaceViewController()
-        navigationController?.pushViewController(nextVC, animated: false)
-    }
 
+    @objc func nextEvent(){
+        let nextVC = RequestResultViewController()
+        navigationController?.pushViewController(nextVC, animated: true)
+    }
+    
     @objc func backTapped(sender: UIBarButtonItem) {
         navigationController?.popViewController(animated: false)
     }
-    
-    func calendarSetting(){
-        
-        fsCalendar.delegate = self
-        fsCalendar.dataSource = self
-        
-        fsCalendar.appearance.weekdayTextColor = .black
-        fsCalendar.appearance.headerTitleColor = .black
+
+    @objc func iWillGoEvent(sender: UIBarButtonItem) {
+        if isIWillGoSelected == false && isComeHereSelected == false {
+            isIWillGoSelected = true
+            iWillGoBtn.setImage(UIImage(named: "selectedIWillGoIcon.svg"), for: .normal)
+        }else{
+            isIWillGoSelected = false
+            iWillGoBtn.setImage(UIImage(named: "iWillGoIcon.svg"), for: .normal)
+        }
     }
-
-}
-
-
-extension PickDateViewController: FSCalendarDataSource, FSCalendarDelegate{
+    
+    @objc func comeHereEvent(sender: UIBarButtonItem) {
+        if isComeHereSelected == false && isIWillGoSelected == false {
+            isComeHereSelected = true
+            comeHereBtn.setImage(UIImage(named: "selectedComeHereIcon.svg"), for: .normal)
+        }else{
+            isComeHereSelected = false
+            comeHereBtn.setImage(UIImage(named: "comeHereIcon.svg"), for: .normal)
+        }
+    }
     
 }
