+++ FITI/FITI/Main/Detail/Request/PickService/PickServiceViewController.swift
@@ -1,5 +1,5 @@
 //
-//  PickPlaceViewController.swift
+//  PickServiceViewController.swift
 //  FITI
 //
 //  Created by 홍준혁 on 2023/01/04.
@@ -7,17 +7,13 @@
 
 import UIKit
 import SnapKit
-import Then
 
-class PickPlaceViewController: UIViewController {
-    
-    var isIWillGoSelected : Bool = false
-    var isComeHereSelected : Bool = false
+class PickServiceViewController: UIViewController {
 
     var titleLabel : UILabel = {
         let label = UILabel()
-        label.font = UIFont.systemFont(ofSize: 20.0)
-        label.text = "픽업 형태"
+        label.font = UIFont(name: "Avenir-Black", size: 20.0)
+        label.text = "서비스 선택"
         label.textColor = UIColor.customColor(.blue)
         return label
     }()
@@ -39,30 +35,14 @@ class PickPlaceViewController: UIViewController {
         }
         return view
     }()
-
-    lazy var iWillGoBtn : UIButton = {
-        let btn = UIButton()
-        btn.setImage(UIImage(named: "iWillGoIcon.svg"), for: .normal)
-        btn.addTarget(self, action: #selector(iWillGoEvent), for: .touchUpInside)
-        return btn
-    }()
     
-    lazy var comeHereBtn : UIButton = {
-        let btn = UIButton()
-        btn.setImage(UIImage(named: "comeHereIcon.svg"), for: .normal)
-        btn.addTarget(self, action: #selector(comeHereEvent), for: .touchUpInside)
-        return btn
+    var subTitleLabel : UILabel = {
+        let label = UILabel()
+        label.font = UIFont.systemFont(ofSize: 20.0)
+        label.text = "서비스 선택하기"
+        label.textColor = UIColor.black
+        return label
     }()
-    
-    lazy var globalStackView : UIStackView = {
-        let stackView = UIStackView(arrangedSubviews: [iWillGoBtn,comeHereBtn])
-        stackView.axis = .horizontal
-        stackView.spacing = 20
-        stackView.alignment = .bottom
-        stackView.distribution = .fillEqually
-        return stackView
-    }()
-    
     
     private let nextBtn : UIButton = {
         let btn = UIButton()
@@ -76,7 +56,8 @@ class PickPlaceViewController: UIViewController {
         btn.addTarget(self, action: #selector(nextEvent), for: .touchUpInside)
         return btn
     }()
-    
+
+    var pickStackView : UIView = PickServiceView()
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -96,8 +77,9 @@ class PickPlaceViewController: UIViewController {
         view.addSubview(titleLabel)
         view.addSubview(progressView)
         progressView.addSubview(grayView)
-        view.addSubview(globalStackView)
+        view.addSubview(subTitleLabel)
         view.addSubview(nextBtn)
+        view.addSubview(pickStackView)
     }
     
     private func setConstraints(){
@@ -111,14 +93,19 @@ class PickPlaceViewController: UIViewController {
             make.trailing.equalToSuperview()
         }
         grayView.snp.makeConstraints { make in
-            make.trailing.equalToSuperview()
+            make.leading.equalToSuperview()
             make.width.equalToSuperview().dividedBy(3)
         }
-        globalStackView.snp.makeConstraints { make in
-            make.top.equalTo(progressView.snp.bottom).offset(65)
-            make.leading.equalToSuperview().offset(20)
-            make.trailing.equalToSuperview().offset(-20)
-            make.bottom.equalToSuperview().offset(-430)
+        subTitleLabel.snp.makeConstraints { make in
+            make.top.equalTo(grayView.snp.bottom).offset(33)
+            make.leading.equalToSuperview().offset(30)
+            make.trailing.equalToSuperview().offset(-220)
+        }
+        pickStackView.snp.makeConstraints { make in
+            make.top.equalTo(subTitleLabel.snp.bottom).offset(40)
+            make.leading.equalToSuperview().offset(30)
+            make.trailing.equalToSuperview().offset(-30)
+            make.height.equalTo(80)
         }
         nextBtn.snp.makeConstraints { make in
             make.top.equalToSuperview().offset(770)
@@ -127,34 +114,14 @@ class PickPlaceViewController: UIViewController {
             make.bottom.equalToSuperview().offset(-40)
         }
     }
-
+    
     @objc func nextEvent(){
-        let nextVC = RequestResultViewController()
-        navigationController?.pushViewController(nextVC, animated: true)
+        let nextVC = PickDateViewController()
+        navigationController?.pushViewController(nextVC, animated: false)
     }
     
     @objc func backTapped(sender: UIBarButtonItem) {
-        navigationController?.popViewController(animated: false)
+        navigationController?.popViewController(animated: true)
     }
 
-    @objc func iWillGoEvent(sender: UIBarButtonItem) {
-        if isIWillGoSelected == false && isComeHereSelected == false {
-            isIWillGoSelected = true
-            iWillGoBtn.setImage(UIImage(named: "selectedIWillGoIcon.svg"), for: .normal)
-        }else{
-            isIWillGoSelected = false
-            iWillGoBtn.setImage(UIImage(named: "iWillGoIcon.svg"), for: .normal)
-        }
-    }
-    
-    @objc func comeHereEvent(sender: UIBarButtonItem) {
-        if isComeHereSelected == false && isIWillGoSelected == false {
-            isComeHereSelected = true
-            comeHereBtn.setImage(UIImage(named: "selectedComeHereIcon.svg"), for: .normal)
-        }else{
-            isComeHereSelected = false
-            comeHereBtn.setImage(UIImage(named: "comeHereIcon.svg"), for: .normal)
-        }
-    }
-    
 }
